Extract wire protocol and axum glue into http-wire crate (#4780)
* refactor(wire): extract tide-disco wire protocol into http-wire crate

The protocol tide-disco defined (Accept/Content-Type negotiation between
vbs binary and JSON, the {status, message} error envelope, healthcheck
types, per-content-type WebSocket frame formats) was implemented twice:
once in espresso-api's axum layer and once, mirror-image, in http-client.
The two copies had already diverged (the server tolerated content-type
parameters, the client matched exactly) and kept three copies of the
health types and negotiation logic in sync only by parallel regression
tests.

http-wire is now the single implementation. espresso-api keeps only the
axum glue (WireFormat, respond, healthcheck responses, drive_ws_stream)
and http-client only the reqwest/tungstenite transport shell; both call
the same codecs, so wire compatibility holds by construction. The
load-bearing tide-disco compatibility tests (variant ordinals, the
Unavailabale misspelling, the numeric status envelope) move next to the
one implementation they constrain.

Client-visible behavior is unchanged except that response content types
are now matched by media-type essence (parameters and case tolerated),
the same tolerance the servers adopted in #4704.

* move more stuff

* refactor(http-wire): rename crate and gate axum glue behind a feature

Renames espresso-wire back to http-wire, pairing with http-client.

The server glue (WireFormat, respond, healthcheck responses,
drive_ws_stream, cors_layer) now sits behind a `server` feature so a
pure reqwest/tungstenite client no longer compiles axum, tower-http and
tokio for code it can never call. Measured: http-client's dependency
tree now has zero axum, and the five crates still on surf-disco that
migrate next (hotshot-task-impls, light-client, client,
hotshot-events-service, espresso-contract-deployer) have no axum today
and would otherwise have gained one.

Cargo cannot override a workspace dependency's default features from a
member, so servers opt in with features = ["server"] instead of the
client opting out. The crate keeps default = ["server"] so a standalone
`cargo test -p http-wire` still runs the server tests.
