+++ crates/espresso/api/Cargo.toml
@@ -11,6 +11,7 @@ axum = { workspace = true }
 base64 = { workspace = true }
 futures = { workspace = true }
 hex = { workspace = true }
+http-wire = { workspace = true, features = ["server"] }
 prost = { workspace = true }
 schemars = { workspace = true }
 serde = { workspace = true }
@@ -22,7 +23,6 @@ tonic = { workspace = true }
 tonic-prost = { workspace = true }
 tonic-reflection = { workspace = true }
 tower = { workspace = true, features = ["limit", "load-shed", "util"] }
-tower-http = { workspace = true, features = ["cors"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 url = { workspace = true }
