+++ crates/espresso/node/src/bin/submit-transactions.rs
@@ -14,7 +14,6 @@ use clap::Parser;
 use committable::{Commitment, Committable};
 #[cfg(feature = "benchmarking")]
 use csv::Writer;
-use espresso_api::{cors_layer, healthcheck_response};
 use espresso_node::SequencerApiVersion;
 use espresso_types::{SeqTypes, Transaction, parse_duration, parse_size};
 use espresso_utils::logging;
@@ -24,6 +23,7 @@ use futures::{
     stream::StreamExt,
 };
 use hotshot_query_service::{Error, availability::BlockQueryData, types::HeightIndexed};
+use http_wire::{cors_layer, healthcheck_response};
 use rand::{Rng, RngCore, SeedableRng};
 use rand_chacha::ChaChaRng;
 use rand_distr::Distribution;
