+++ crates/hotshot/orchestrator/src/lib.rs
@@ -30,10 +30,6 @@ use axum::{
 };
 use client::{BenchResults, BenchResultsDownloadConfig};
 use csv::Writer;
-use espresso_api::{
-    cors_layer, healthcheck_response,
-    wire::{self, WireFormat},
-};
 use futures::{StreamExt, stream::FuturesUnordered};
 use hotshot_types::{
     PeerConfig,
@@ -43,6 +39,7 @@ use hotshot_types::{
         signature_key::{SignatureKey, StakeTableEntryType},
     },
 };
+use http_wire::{self as wire, WireFormat, cors_layer, healthcheck_response};
 use libp2p_identity::{
     Keypair, PeerId,
     ed25519::{Keypair as EdKeypair, SecretKey},
