+++ crates/http-client/Cargo.toml
@@ -4,14 +4,14 @@ version = { workspace = true }
 authors = { workspace = true }
 edition = { workspace = true }
 description = "Reqwest-based HTTP/WebSocket client for tide-disco-shaped APIs."
+publish = false
 
 [dependencies]
 futures = { workspace = true }
 hex = { workspace = true }
+http-wire = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
-serde_json = { workspace = true }
-thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-tungstenite = { workspace = true, features = ["rustls-tls-webpki-roots"] }
 tracing = { workspace = true }
