+++ crates/http-wire/src/server.rs
@@ -1,0 +1,280 @@
+//! Axum glue for serving the wire protocol.
+//!
+//! Everything a service needs regardless of its routes: content-negotiated responses,
+//! request-body decoding, the healthcheck response shapes, the permissive CORS layer, and one
+//! correct way to drive a WebSocket stream endpoint. Services describe their own error envelope
+//! with a [`WireFormat`] impl and share everything else here, so none of them needs to depend on
+//! another service's API crate.
+
+use axum::{
+    Json,
+    extract::ws::{Message, WebSocket},
+    http::{HeaderMap, StatusCode, header},
+    response::{IntoResponse, Response},
+};
+use futures::{StreamExt, stream::BoxStream};
+use serde::{Serialize, de::DeserializeOwned};
+use tower_http::cors::{Any, CorsLayer};
+use vbs::version::{StaticVersion, StaticVersionType};
+
+use crate::{
+    body::{DecodeFailure, encode_body},
+    content_type::ContentType,
+    health::{AppHealth, HealthStatus},
+    ws::{encode_binary_frame, encode_text_frame},
+};
+
+/// Describes a service's wire protocol: its VBS framing version and its error envelope.
+pub trait WireFormat {
+    /// The service's wire error envelope, serialized in the negotiated format like any body.
+    type Error: Serialize;
+    /// VBS framing version for binary bodies.
+    type Version: StaticVersionType;
+
+    /// The HTTP status an error response is sent with.
+    fn status(err: &Self::Error) -> StatusCode;
+
+    /// The error reported when a response body fails to serialize. The message names the
+    /// format that failed.
+    fn serialize_failure(message: String) -> Self::Error;
+}
+
+/// Encode a successful response body, negotiating VBS binary vs JSON from the `Accept` header.
+pub fn encode_ok<C: WireFormat, T: Serialize>(headers: &HeaderMap, value: T) -> Response {
+    let format = ContentType::negotiate(headers);
+    match encode_body::<C::Version, _>(format, &value) {
+        Ok(bytes) => ([(header::CONTENT_TYPE, format.mime())], bytes).into_response(),
+        Err(err) => encode_err::<C>(headers, C::serialize_failure(err.to_string())),
+    }
+}
+
+/// Encode an error response using the same content negotiation as [`encode_ok`].
+pub fn encode_err<C: WireFormat>(headers: &HeaderMap, err: C::Error) -> Response {
+    let status = C::status(&err);
+    let format = ContentType::negotiate(headers);
+    match encode_body::<C::Version, _>(format, &err) {
+        Ok(bytes) => (status, [(header::CONTENT_TYPE, format.mime())], bytes).into_response(),
+        // The envelope itself failed to serialize; fall back to JSON so the client at least
+        // sees the status and a body (axum's `Json` degrades to a plain 500 if even that fails).
+        Err(_) => (status, Json(err)).into_response(),
+    }
+}
+
+pub fn respond<C: WireFormat, T: Serialize>(
+    headers: &HeaderMap,
+    result: Result<T, C::Error>,
+) -> Response {
+    match result {
+        Ok(value) => encode_ok::<C, _>(headers, value),
+        Err(err) => encode_err::<C>(headers, err),
+    }
+}
+
+/// Decode a request body: VBS for `application/octet-stream`, JSON for `application/json`.
+/// Content types are matched by media-type essence so parameters (e.g. a charset) are tolerated.
+pub fn decode_body<Ver: StaticVersionType, T: DeserializeOwned>(
+    headers: &HeaderMap,
+    body: &[u8],
+) -> Result<T, DecodeFailure> {
+    let content_type = headers
+        .get(header::CONTENT_TYPE)
+        .and_then(|v| v.to_str().ok());
+    crate::body::decode_body::<Ver, T>(content_type, body)
+}
+
+/// App-level `/healthcheck` response, in JSON or vbs binary depending on `Accept`.
+///
+/// Reports [`AppHealth`] with no per-module entries, the shape every service with at least one
+/// registered module has always served.
+pub fn healthcheck_response(headers: &HeaderMap) -> Response {
+    encode_health(headers, &AppHealth::default())
+}
+
+/// Per-module `/healthcheck` response: a bare [`HealthStatus`], the shape served for a module
+/// without an explicit healthcheck handler.
+pub fn module_healthcheck_response(headers: &HeaderMap) -> Response {
+    encode_health(headers, &HealthStatus::Available)
+}
+
+/// Permissive CORS: allow any origin, method, and headers on every response.
+///
+/// Apply it as the outermost layer, so it also covers responses generated by middleware rather
+/// than a handler (a connection limit's 429 would otherwise skip it).
+///
+/// `allow_methods` and `allow_headers` are only echoed on the `OPTIONS` preflight, which is where
+/// a browser reads them; `allow_origin` goes out on every response.
+pub fn cors_layer() -> CorsLayer {
+    CorsLayer::new()
+        .allow_methods(Any)
+        .allow_headers(Any)
+        .allow_origin(Any)
+}
+
+/// Forwards `stream` over `socket` in the format negotiated by [`ContentType::negotiate`] until
+/// the stream ends or the client disconnects, then performs the close handshake. Production
+/// stream consumers default to `Accept: application/octet-stream` and expect VBS-encoded
+/// `Message::Binary` frames.
+///
+/// This is the one shared way to serve a socket stream endpoint; hand-rolled write-only loops
+/// leak the task and its data-source subscription whenever a client disconnects while the stream
+/// is quiet.
+///
+/// `Ver` is the VBS framing version of binary frames, and must be the same version the service's
+/// [`WireFormat`] uses for its request/response bodies.
+pub async fn drive_ws_stream<Ver: StaticVersionType, T: Serialize>(
+    mut socket: WebSocket,
+    stream: BoxStream<'static, T>,
+    format: ContentType,
+) {
+    futures::pin_mut!(stream);
+    loop {
+        // Also poll the client side: a disconnect must end this task even while the stream is
+        // quiet, or the socket's connection slot and the stream task leak until the next send.
+        let item = tokio::select! {
+            item = stream.next() => item,
+            msg = socket.recv() => match msg {
+                None | Some(Err(_)) | Some(Ok(Message::Close(_))) => return,
+                Some(Ok(_)) => continue,
+            },
+        };
+        let Some(item) = item else { break };
+        let frame = match format {
+            ContentType::Binary => {
+                encode_binary_frame::<Ver, _>(&item).map(|bytes| Message::Binary(bytes.into()))
+            },
+            ContentType::Json => encode_text_frame(&item).map(|json| Message::Text(json.into())),
+        };
+        let Ok(msg) = frame else { break };
+        if socket.send(msg).await.is_err() {
+            return;
+        }
+    }
+    // Close handshake, like the socket handlers before the axum migration. Without it, dropping
+    // the socket resets the connection and clients see an error instead of end-of-stream; finite
+    // streams rely on a clean close to signal completion.
+    let _ = socket.send(Message::Close(None)).await;
+}
+
+/// All services frame binary healthcheck responses with v0.1; the shape predates per-service
+/// versioning.
+type HealthcheckVersion = StaticVersion<0, 1>;
+
+fn encode_health<T: Serialize>(headers: &HeaderMap, value: &T) -> Response {
+    let format = ContentType::negotiate(headers);
+    match encode_body::<HealthcheckVersion, _>(format, value) {
+        Ok(bytes) => ([(header::CONTENT_TYPE, format.mime())], bytes).into_response(),
+        Err(err) => (StatusCode::INTERNAL_SERVER_ERROR, err.to_string()).into_response(),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use axum::http::HeaderValue;
+    use serde::Deserialize;
+    use vbs::{BinarySerializer, Serializer};
+
+    use super::*;
+
+    #[derive(Debug, Serialize, Deserialize, PartialEq)]
+    struct TestError {
+        status: u16,
+        message: String,
+    }
+
+    struct TestFormat;
+
+    impl WireFormat for TestFormat {
+        type Error = TestError;
+        type Version = StaticVersion<0, 1>;
+
+        fn status(err: &Self::Error) -> StatusCode {
+            StatusCode::from_u16(err.status).unwrap_or(StatusCode::INTERNAL_SERVER_ERROR)
+        }
+
+        fn serialize_failure(message: String) -> Self::Error {
+            TestError {
+                status: 500,
+                message,
+            }
+        }
+    }
+
+    fn headers(name: header::HeaderName, value: &str) -> HeaderMap {
+        let mut headers = HeaderMap::new();
+        headers.insert(name, HeaderValue::from_str(value).unwrap());
+        headers
+    }
+
+    async fn body_bytes(resp: Response) -> Vec<u8> {
+        axum::body::to_bytes(resp.into_body(), usize::MAX)
+            .await
+            .unwrap()
+            .to_vec()
+    }
+
+    #[tokio::test]
+    async fn encode_negotiates_binary_and_json() {
+        let binary = headers(header::ACCEPT, "application/octet-stream");
+        let resp = encode_ok::<TestFormat, _>(&binary, 42u64);
+        assert_eq!(resp.status(), StatusCode::OK);
+        assert_eq!(
+            resp.headers().get(header::CONTENT_TYPE).unwrap(),
+            "application/octet-stream"
+        );
+        let bytes = body_bytes(resp).await;
+        let decoded: u64 = Serializer::<StaticVersion<0, 1>>::deserialize(&bytes).unwrap();
+        assert_eq!(decoded, 42);
+
+        let resp = encode_ok::<TestFormat, _>(&HeaderMap::new(), 42u64);
+        assert_eq!(resp.status(), StatusCode::OK);
+        assert_eq!(String::from_utf8(body_bytes(resp).await).unwrap(), "42");
+    }
+
+    #[tokio::test]
+    async fn errors_carry_status_and_envelope() {
+        let err = TestError {
+            status: 404,
+            message: "no such thing".into(),
+        };
+        let resp = encode_err::<TestFormat>(&HeaderMap::new(), err);
+        assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+        let decoded: TestError = serde_json::from_slice(&body_bytes(resp).await).unwrap();
+        assert_eq!(decoded.message, "no such thing");
+
+        let binary = headers(header::ACCEPT, "application/octet-stream");
+        let err = TestError {
+            status: 404,
+            message: "no such thing".into(),
+        };
+        let resp = encode_err::<TestFormat>(&binary, err);
+        assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+        let decoded: TestError =
+            Serializer::<StaticVersion<0, 1>>::deserialize(&body_bytes(resp).await).unwrap();
+        assert_eq!(decoded.status, 404);
+    }
+
+    #[test]
+    fn decode_matches_content_type_by_essence() {
+        let json = headers(header::CONTENT_TYPE, "application/json; charset=utf-8");
+        let decoded: u64 = decode_body::<StaticVersion<0, 1>, _>(&json, b"42").unwrap();
+        assert_eq!(decoded, 42);
+
+        let binary = headers(header::CONTENT_TYPE, "application/octet-stream");
+        let bytes = Serializer::<StaticVersion<0, 1>>::serialize(&42u64).unwrap();
+        let decoded: u64 = decode_body::<StaticVersion<0, 1>, _>(&binary, &bytes).unwrap();
+        assert_eq!(decoded, 42);
+
+        assert!(matches!(
+            decode_body::<StaticVersion<0, 1>, u64>(&HeaderMap::new(), b"42"),
+            Err(DecodeFailure::UnsupportedContentType)
+        ));
+        assert!(matches!(
+            decode_body::<StaticVersion<0, 1>, u64>(&json, b"not json"),
+            Err(DecodeFailure::Json(_))
+        ));
+        assert!(matches!(
+            decode_body::<StaticVersion<0, 1>, u64>(&binary, b""),
+            Err(DecodeFailure::Binary(_))
+        ));
+    }
+}
