+++ node-metrics/Cargo.toml
@@ -18,13 +18,13 @@ bincode = { workspace = true }
 bitvec = { workspace = true }
 circular-buffer = { workspace = true }
 clap = { workspace = true }
-espresso-api = { workspace = true }
 espresso-types = { workspace = true, features = ["node"] }
 espresso-utils = { workspace = true, features = ["full"] }
 futures = { workspace = true }
 hotshot = { workspace = true }
 hotshot-example-types = { workspace = true }
 hotshot-query-service = { workspace = true }
+http-wire = { workspace = true, features = ["server"] }
 
 # Dependencies for feature `testing`
 hotshot-types = { workspace = true }
